Set the CODEQL_WRAPPER env var
The CODEQL_WRAPPER environment variable will be set to "action" if
running in an actions context and "runner" otherwise. This variable
will be used in the cli to determine which context it is in.

Also, use a similar mechanism to provide a different user agent
when communicating with the github api.