pass CODEQL_FEATURE_RAM_2023 to the CLI
This prepares for an upcoming change in how the `--ram` argument to
the CLI is interpreted.

The CLI will interpret _either_ this feature flag _or_ an Action
version of 2.22.10 or earlier as a sign that it would fall back to
the current interpretation of `--ram`.

Later in the future, the Action can start using a less conservative
calculation of `--ram` (when it detect the CLI is sufficiently new
CLIs) and signal that to the CLI by _not_ passing
`CODEQL_FEATURE_RAM_2023` anyway.

That way we can avoid wrong RAM settings caused by mismatched CLI and
Actions versions.