Switch `debug-artifacts` job back to `ubuntu`
Now that CLI v2.17.4+ are available, we can switch this job back to `ubuntu`. As a result, we can also bring back testing on the older CLI versions (which did not work on MacOS).
CLI v.2.16.6 has a known failure on Linux so we exclude it from this workflow.

This change is orthogonal to the PR check generator change as the check doesn't use the generator.